Break out Custom LED to separate page
