+++ src/Components/Navigation.js
@@ -47,6 +47,7 @@ const Navigation = (props) => {
 						<NavDropdown.Item as={NavLink} exact="true" to="/pin-mapping">Pin Mapping</NavDropdown.Item>
 						<NavDropdown.Item as={NavLink} exact="true" to="/keyboard-mapping">Keyboard Mapping</NavDropdown.Item>
 						<NavDropdown.Item as={NavLink} exact="true" to="/led-config">LED Configuration</NavDropdown.Item>
+						<NavDropdown.Item as={NavLink} exact="true" to="/custom-theme">Custom LED Theme</NavDropdown.Item>
 						<NavDropdown.Item as={NavLink} exact="true" to="/display-config">Display Configuration</NavDropdown.Item>
 						<NavDropdown.Item as={NavLink} exact="true" to="/add-ons">Add-Ons Configuration</NavDropdown.Item>
 						<NavDropdown.Item as={NavLink} exact="true" to="/backup">Data Backup and Restoration</NavDropdown.Item>
